+++ CHANGELOG.md
@@ -17,6 +17,7 @@
   - Smart check all available refers to require, adding refer options to `Add require` code actions. #627
   - Add setting `keep-parens-when-threading?` to keep parens for single arity functions when threading. #636
   - Avoid adding duplicate requires when adding a new require via code action. #640
+  - Improve common known snippets to replace completion items, improving completion UX.
 
 This release was supported by [Clojurists Together](https://www.clojuriststogether.org/)